cleaning up
Removed commented code, moved constants, and fixed theme for ConnectKit provider.

Console errors need to be resolved by library creators adding support for Fantom officially.
