+++ src/components/Header/index.tsx
@@ -12,7 +12,6 @@ import NavLinks from './NavLinks';
 import { toggleMenu, togglePopup } from '../../redux/navigation/slice';
 
 const Header = () => {
-  // const [isOpen, setIsOpen] = useState(false);
   const dispatch = useDispatch();
   const openWallet = useSelector(selectWalletOpen);
   const openMenu = useSelector(selectMenuOpen);
@@ -24,7 +23,6 @@ const Header = () => {
     }
   };
 
-  // const handleWalletClick = () => dispatch(togglePopup({ value: !openWallet }));
   const handleMenuClose = () => dispatch(toggleMenu({ value: false }));
 
   return (
@@ -51,47 +49,7 @@ const Header = () => {
 
             {/* Using ConnectKit wallet button */}
             <ConnectKitButton />
-
-            {/* <Link href="#">
-              <div
-                role="none"
-                className="button user"
-                id="popupBtn"
-                onClick={handleWalletClick}
-              >
-                <span>
-                  <img src="/images/icons/metamask.svg" alt="" />
-                </span>
-                <p>0xC6c...B77D</p>
-              </div>
-            </Link> */}
           </div>
-
-          {/* Wallet popup (old):
-           <div className={`popup ${openWallet ? 'active' : ''}`} id="popUp">
-            <h4>Wallet</h4>
-            <div className="input">
-              <input type="text" placeholder="0xC6c...B77D" />
-              <button type="button" className="input__button">
-                <img src="/images/icons/copy.svg" alt="" />
-              </button>
-              <button type="button" className="input__button">
-                <img src="/images/icons/share.svg" alt="" />
-              </button>
-            </div>
-            <div className="popup__row">
-              <h5>
-                <span>
-                  <img src="/images/icons/icon-logo.svg" alt="" />
-                </span>
-                MRX
-              </h5>
-              <p>0</p>
-            </div>
-            <button type="button" className="button">
-              Disconnect
-            </button>
-          </div> */}
           <div
             role="none"
             className="burger"
