+++ src/pages/_app.tsx
@@ -7,42 +7,20 @@ import superjson from 'superjson';
 import { Toaster } from 'react-hot-toast';
 import dynamic from 'next/dynamic';
 import Head from 'next/head';
-import { WagmiConfig, createClient, Chain } from 'wagmi';
+import { WagmiConfig, createClient } from 'wagmi';
 import { ConnectKitProvider, getDefaultClient } from 'connectkit';
 
 import '../scss/main.scss';
-import { URL } from '../constants';
+import { URL, ftmChain } from '../constants';
 import { AppRouter } from '../server/route/app.router';
 import { store } from '../redux/store';
 import Layout from '../components/Layout';
-
-const ftmScan = { name: 'Ftmscan', url: 'https://ftmscan.com/' }; // blockexplorer for Fantom
-
-// Since Fantom is not a default chain supported by wagmi, we need to import the Chain type and create our own
-const ftm: Chain = {
-  id: 250,
-  name: 'Fantom',
-  network: 'Fantom Opera',
-  nativeCurrency: { name: 'Fantom', symbol: 'FTM', decimals: 18 },
-  rpcUrls: {
-    default: 'https://rpc.ftm.tools/',
-    public: 'https://rpc.ftm.tools/',
-  },
-  blockExplorers: {
-    etherscan: ftmScan,
-    default: ftmScan,
-  },
-  multicall: {
-    address: '0xb828C456600857abd4ed6C32FAcc607bD0464F4F',
-    blockCreated: 2323970,
-  },
-};
 
 // Creating the web3 client in wagmi, using "getDefaultClient" from ConnectKit which simplifies for us.
 const client = createClient(
   getDefaultClient({
     appName: 'Morphex',
-    chains: [ftm],
+    chains: [ftmChain],
   })
 );
 
@@ -52,7 +30,7 @@ const client = createClient(
 function MyApp({ Component, pageProps }: AppProps) {
   return (
     <WagmiConfig client={client}>
-      <ConnectKitProvider>
+      <ConnectKitProvider theme="auto" mode="light">
         <Provider store={store}>
           <Head>
             <title>Morphex</title>
